requirements.txt
wheel
setuptools
dbt-snowflake
modal
GitPython